@@ -24,30 +24,33 @@
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview",
-    "check": "svelte-check --tsconfig ./tsconfig.app.json && tsgo -p tsconfig.node.json"
+    "check": "svelte-check --tsconfig ./tsconfig.app.json && tsgo -p tsconfig.node.json",
+    "create-module": "node scripts/create-module.cjs",
+    "remove-module": "node scripts/remove-module.cjs"
   },
   "dependencies": {
-    "@internationalized/date": "^3.10.0",
-    "bits-ui": "^2.14.4",
+    "@internationalized/date": "^3.10.1",
+    "bits-ui": "^2.15.4",
+    "jsencrypt": "^3.5.4",
     "primeicons": "^7.0.0",
     "tailwindcss-primeui": "^0.6.1"
   },
   "devDependencies": {
     "@sveltejs/vite-plugin-svelte": "^6.2.1",
-    "@tailwindcss/postcss": "^4.1.17",
+    "@tailwindcss/postcss": "^4.1.18",
     "@tsconfig/svelte": "^5.0.6",
-    "@types/node": "^24.10.1",
-    "@typescript/native-preview": "7.0.0-dev.20251207.1",
-    "autoprefixer": "^10.4.22",
+    "@types/node": "^24.10.4",
+    "@typescript/native-preview": "7.0.0-dev.20260105.1",
+    "autoprefixer": "^10.4.23",
     "flowbite": "^4.0.1",
-    "flowbite-svelte": "^1.28.3",
+    "flowbite-svelte": "^1.31.0",
     "postcss": "^8.5.6",
-    "svelte": "^5.43.8",
-    "svelte-check": "^4.3.4",
-    "tailwindcss": "^4.1.17",
-    "vite": "^7.2.7"
+    "svelte": "^5.46.1",
+    "svelte-check": "^4.3.5",
+    "tailwindcss": "^4.1.18",
+    "vite": "^7.3.0"
   },
   "optionalDependencies": {
-    "@rollup/rollup-win32-x64-msvc": "^4.53.3"
+    "@rollup/rollup-win32-x64-msvc": "^4.55.1"
   }
 }
